@@ -1,15 +1,16 @@
 name: http_exception
-version: 0.2.1
+version: 0.2.2
 description: A set of HTTP exceptions.
 homepage: https://github.com/bwu-dart/http_exception
 authors:
   - Günter Zöchbauer <guenter@gzoechbauer.com>
   - Thomas Profelt <tom@octobyte.at>
+  - Diego Perez <eazel7@gmail.com>
 environment:
-  sdk: '>=2.0.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
 
 dev_dependencies:
-  grinder: ^0.8.3
-  test: ^1.3.0
+  grinder: ^0.9.0
+  test: ^1.19.5
